pyproject.toml
[tool.poetry]
name = "sigma"
version = "0.1.0"
description = "Python library for parsing and interacting with Sigma rules."
authors = ["Caleb Stewart <caleb.stewart94@gmail.com>"]

[tool.poetry.scripts]
sigma = "sigma.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
pyparser = "^1.0"
PyYAML = "^6.0"
rich = "^11.1.0"
sphinx-toolbox = { version = "^2.15.2", optional = true }
Sphinx = { version= "^4.0.2", optional = true }
enum-tools = { version= "^0.7.0", optional = true }
furo = { version= "^2021.11.23", optional = true }
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.8.0"
flake8 = "^3.9.2"

[tool.poetry.extras]
docs = ["sphinx-toolbox", "Sphinx", "enum-tools", "furo"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
